requirements.txt
async-dramatiq
redis
fastapi[standard]

aiohttp